% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_tibble.R
\name{to_tibble}
\alias{to_tibble}
\title{to_tibble}
\usage{
to_tibble(object, valuename = "value")
}
\arguments{
\item{object}{list output of fileply}

\item{valuename}{name of the value list-column}
}
\value{
an object of class tibble
}
\description{
convert list output of fileply to an object of class tibble
}
\examples{
# split-apply-combine
write.table(mtcars, "mtcars.csv", row.names = FALSE, sep = ",")
temp <- fileply(file     = "mtcars.csv"
             , groupby = c("carb", "gear")
             , fun     = identity
             , collect = "list"
             , sep     =  ","
             , header  = TRUE
             )
temp
to_tibble(temp)
unlink("mtcars.csv")


}

